agents/schema-agent/requirements.txt
sentence-transformers
faiss-cpu
einops
llama-cpp-python
ollama
